+++ src/minishell1.c
@@ -1,5 +1,33 @@
 #include "minishell.h"
 
+/**
+ * * Disables CTRL hotkey(+c) from printing ^C 
+*/
+static void	disable_ctrl_c_hotkey(void)
+{
+	struct termios	termios;
+	int				rc;
+
+	rc = tcgetattr(0, &termios);
+	if (rc)
+	{
+		perror("tcgetattr");
+		exit(1);
+	}
+	termios.c_lflag &= ~ECHOCTL;
+	rc = tcsetattr(0, 0, &termios);
+	if (rc)
+	{
+		perror("tcsetattr");
+		exit(1);
+	}
+}
+
+/**
+ * * Process every input line sended on STDIN_FILENO
+ * @param list	list
+ * @param str	new str to link
+*/
 static void	processline(void)
 {
 	char	*str_got;
@@ -30,6 +58,7 @@ int	main(int argc, char **argv, char **env)
 	(void)argv;
 	g_var.env = NULL;
 	g_var.last_cmd_status = 0;
+	disable_ctrl_c_hotkey();
 	init_ms_env(env, &g_var.env);
 	while (1)
 	{