+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2021-12-24 - Edu
+### Added
+- **minishell1.c** Added **disable_ctrl_c_hotkey** to disable CTRL hotkey(+c) from printing ^C following 
+	+ https://unix.stackexchange.com/questions/558694/why-do-terminals-sometimes-echo-special-characters-like-c
+	+ https://stackoverflow.com/questions/68602211/forcing-a-terminal-not-to-print-ctrl-hotkeys-when-signals-are-caught
+### Changed
+- Remove unnecessary code to handle this behaviour on signal_handler and constants.h
+
 ## 2021-12-20 - Edu
 ### Fixed
 - **var_expansor.c** Fix expansions at start of string